Upd: refactoring done
